resolved comments

+++ src/utils.rs
@@ -13,10 +13,10 @@ pub fn date_from(date_string: &str) -> NaiveDate {
     NaiveDate::from_ymd_opt(year, month, day).unwrap()
 }
 
-pub fn parse_date_from(date_str: &str) -> (i32, i32) {
+pub fn parse_date_from(date_str: &str) -> (u16, u16) {
     let mut split_date = date_str.split("/");
     (
-        split_date.next().unwrap().parse::<i32>().unwrap(),
-        split_date.next().unwrap().parse::<i32>().unwrap(),
+        split_date.next().unwrap().parse::<u16>().unwrap(),
+        split_date.next().unwrap().parse::<u16>().unwrap(),
     )
 }
